[refactor][state/changelog] Extract StateChangeUploadScheduler interface from StateChangeUploader
Motivation: the two interfaces should have different semantics
regarding returning upload result vs completing upload tasks with it calling upload().
That would allow to discard unnecessarily uploaded state in subsequent commit.
